Finishing the homework, removing all errors
