+++ src/views.py
@@ -1,9 +1,12 @@
+import json
 import logging
 import os
 from datetime import date, datetime
-from typing import Any, Dict, Hashable, List
+from typing import Dict
 
-import pandas as pd
+
+from src.utils import (card_info, convert_currency, greeting, read_transactions_exl, read_transactions_exl_all,
+                       result_ticker, sort_by_amount)
 
 current_dir = os.path.dirname(os.path.abspath(__file__))
 project_root = os.path.abspath(os.path.join(current_dir, ".."))
@@ -22,115 +25,29 @@ file_handler.setFormatter(file_formatter)
 logger.addHandler(file_handler)
 
 
-def read_transactions_exl(file_path: str) -> List[Dict[Hashable, Any]]:
-    """Функция получения, чтения файла excel, преобразование в список словарей без нулевых значений по номеру карты"""
-    logger.info(
-        "Запуск функции получения, чтения файла excel, преобразование в список словарей"
-        " без нулевых значений по номеру карты"
-    )
-    df = pd.read_excel(str(file_path))
-    df_filtered = df.dropna(subset=["Номер карты"])
-    list_transactions_exl = df_filtered.to_dict(orient="records")
-    return list_transactions_exl
+def main_page(date_time: str) -> str:
+    logger.info("Запуск функции главной страницы ")
+    with open(data_file_path_json, "r") as f:
+        user_settings = json.load(f)
 
+    dt_obj = datetime.strptime(date_time, "%Y-%m-%d %H:%M:%S")
+    start_of_month = dt_obj.replace(day=1)
+    start_date = start_of_month.strftime("%d.%m.%Y")
+    end_date = dt_obj.strftime("%d.%m.%Y")
 
-def read_transactions_exl_all(file_path: str) -> List[Dict[Hashable, Any]]:
-    """Функция получения, чтения файла excel, преобразование в список словарей  без нулевых значений по дате"""
-    logger.info(
-        "Запуск функции получения, чтения файла excel,"
-        " преобразование в список словарей  без нулевых значений по дате"
-    )
-    df = pd.read_excel(str(file_path))
-    df_filtered = df.dropna(subset=["Дата платежа"])
-    list_transactions_exl_all = df_filtered.to_dict(orient="records")
-    return list_transactions_exl_all
+    transactions = read_transactions_exl(data_file_path_exl)
+    transactions_all = read_transactions_exl_all(data_file_path_exl)
 
+    final_result = {
+        "greeting": greeting(date_time),
+        "cards": [card_info(transactions, start_date, end_date)],
+        "top_transactions": [sort_by_amount(transactions_all, start_date, end_date)],
+        "currency_rates": [convert_currency(user_settings)],
+        "stock_prices": result_ticker(user_settings),
+    }
+    logger.info("Печать результата по главной странице")
 
-def greeting(time_str: str) -> str:
-    """Функция приветствия, в зависимости от указанного времени"""
-    logger.info("Запуск функции приветствия, в зависимости от указанного времени ")
-    dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S")
-
-    hour = dt.hour
-
-    if 5 <= hour < 11:
-        message = str(("Доброе утро!"))
-    elif 11 <= hour < 18:
-        message = str(("Добрый день!"))
-    elif 18 <= hour < 23:
-        message = "Добрый вечер!"
-    else:
-        message = "Доброй ночи!"
-
-    logger.info(f"Сформированное приветствие: {message}")
-    return message
-
-
-def card_info(transactions: List[Dict[str, Any]], start: date, end: date) -> List[Dict[str, float]]:
-    """Функция вывода информации по карте последние 4 цифры, траты, кэшбэк"""
-    logger.info("Запуск функции вывода информации по карте последние 4 цифры, траты, кэшбэк ")
-    result = []
-    trans_list = {}
-
-    if isinstance(start, str):
-        start = datetime.strptime(start, "%d.%m.%Y").date()
-    if isinstance(end, str):
-        end = datetime.strptime(end, "%d.%m.%Y").date()
-
-    for trans in transactions:
-        transaction_date = datetime.strptime(trans["Дата платежа"], "%d.%m.%Y").date()
-        if start <= transaction_date <= end:
-            card_number_f = trans["Номер карты"]
-            card_number = card_number_f[1:5].replace(card_number_f[:-4], card_number_f[-4:])
-            amount = trans.get("Сумма операции", 0)
-
-            if float(amount) < 0:
-
-                if card_number not in trans_list:
-                    trans_list[card_number] = 0
-
-                trans_list[card_number] += int(float(amount))
-
-    for key, value in trans_list.items():
-        result.append(
-            {"last_digits": key, "total_spent": -round(value, 2), "cashback": -round((round(value, 2) / 100), 2)}
-        )
-    logger.info("Вывод результата по картам")
-    return result
-
-
-def sort_by_amount(
-    transactions: List[Dict[str, Any]], start: date, end: date, reverse_str: bool = True
-) -> List[Dict[str, Any]]:
-    """Функция сортировки по тратам по убыванию"""
-    logger.info("Запуск функции сортировки по тратам по убыванию")
-    result = []
-    filtered_transactions = []
-
-    if isinstance(start, str):
-        start = datetime.strptime(start, "%d.%m.%Y").date()
-    if isinstance(end, str):
-        end = datetime.strptime(end, "%d.%m.%Y").date()
-
-    # Фильтруем транзакции по дате
-    for trans in transactions:
-        transaction_date: date = datetime.strptime(trans["Дата платежа"], "%d.%m.%Y").date()
-        if start <= transaction_date <= end:
-            filtered_transactions.append(trans)
-
-    # Сортируем отфильтрованные транзакции по сумме операции
-    sorted_transactions = sorted(filtered_transactions, key=lambda x: float(x["Сумма операции"]), reverse=reverse_str)
-
-    # Формируем результат, ограничивая до 5 записей
-    for i in range(min(len(sorted_transactions), 5)):
-        trans = sorted_transactions[i]
-        result.append(
-            {
-                "date": trans["Дата платежа"],
-                "amount": trans["Сумма операции"],
-                "category": trans["Категория"],
-                "description": trans["Описание"],
-            }
-        )
-    logger.info("Возвращаем результат сортировки")
-    return result
+    ff_result = json.dumps(final_result, indent=4, ensure_ascii=False)
+    with open(data_file_path_result_main_screen, "w", encoding="utf-8") as f:
+        f.write(ff_result)
+    return ff_result
