+++ tests/conftest.py
@@ -11,7 +11,7 @@ def trans_list() -> List[Dict[str, Any]]:
             "Дата платежа": "23.07.2019",
             "Номер карты": "*4556",
             "Статус": "OK",
-            "Категория": "Аптека",
+            "Категория": "Супермаркет",
             "Сумма операции": -5000.0,
             "Валюта операции": "RUB",
             "Сумма платежа": -5000.0,
@@ -22,12 +22,12 @@ def trans_list() -> List[Dict[str, Any]]:
             "Сумма операции с округлением": 5000.0,
         },
         {
-            "Дата операции": "19.07.2019 22:02:30",
-            "Дата платежа": "21.07.2019",
+            "Дата операции": "12.12.2021",
+            "Дата платежа": "12.12.2021",
             "Номер карты": "*7197",
             "Статус": "OK",
-            "Категория": "Аптека",
-            "Сумма операции": -149.0,
+            "Категория": "Аптеки",
+            "Сумма операции": -6486.5,
             "Валюта операции": "RUB",
             "Сумма платежа": -149.0,
             "Валюта платежа": "RUB",
@@ -41,7 +41,7 @@ def trans_list() -> List[Dict[str, Any]]:
             "Дата платежа": "01.07.2019",
             "Номер карты": "*6002",
             "Статус": "FAILED",
-            "Категория": "Аптека",
+            "Категория": "Супермаркет",
             "Сумма операции": -17000.0,
             "Валюта операции": "RUB",
             "Сумма платежа": -17000.0,
@@ -52,3 +52,8 @@ def trans_list() -> List[Dict[str, Any]]:
             "Сумма операции с округлением": 17000.0,
         },
     ]
+
+
+@pytest.fixture
+def date_time()-> str:
+    return "2021-12-12 03:03:03"
